move to row to crc
